narrow-gate/pyproject.toml
[tool.poetry]
name = "narrow-gate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
jupyter = "^1.1.1"
beautifulsoup4 = "^4.12.3"
geopy = "^2.4.1"
googlemaps = "^4.10.0"
plotly = "^5.24.1"
google = "^3.0.0"
python-dotenv = "^1.0.1"
pathlib = "^1.0.1"
streamlit = "^1.40.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
